// src/one4all/airtime/dto/topup.dto.ts
export class TopupDto {
  readonly retailer: string;
  userId: string;
  merchantId?: string;
  recipientNumber: string;
  amount: number;
  charge?: string;
  channel?: number;
  network?: number;
  clientReference?: string;
  transType?: string;
  customerName?: string;
  customerEmail?: string;
  description?: string;
  serviceStatusCode?: string;
  statusNarration?: string;
  paymentStatus?: string;
  redirectURL?: string;
  clientId?: string;
  clientSecret?: string;
  customerMsisdn?: string;
  recipientMsisdn?: string;
  senderMsisdn?: string;
  primaryCallbackUrl?: string;
  merchantName?: string;
  serviceName?: string;
  merchantReference?: string;
  transId: string;
  originalAmount: string;
  amountPaid: string;
  mobileNumber: string;
  recipientName: string;
  senderNumber: string;
  product: string;
  transDescription: string;
  voucher: string;
  serviceStatus: string;
  transMessage: string;
  serviceTransId: string;
  transStatus: string;
  callbackReceipt: string;
  commentary: string;
  callbackUrl: string;
  readonly price: string;
  recipient_number: string;
  sender?: string;
  award?: string;
  apikey: string;
  orderID: string;
  token?: string;
  notifyURL?: string;
  notificationFormat?: string;
  destinationAddress?: string;
  criteria?: string;
  Status?: string;
  transactionId?: string;
  Message?: string;
  MerchantReference?: string;
  CallbackURL?: string;
  currentBalance: string;
  balanceBefore: number;
  balanceAfter: number;
  recipient: string;
  trxn: string;
  fee: any;
  serviceCode: string;
}